+++ src/fs/exfat.h
@@ -88,7 +88,7 @@ public:
 
 //          qint64 minCapacity() const;
     qint64 maxCapacity() const override;
-    qint64 maxLabelLength() const override;
+    int maxLabelLength() const override;
     SupportTool supportToolName() const override;
     bool supportToolFound() const override;
