+++ src/fs/f2fs.h
@@ -93,7 +93,7 @@ public:
 
     qint64 minCapacity() const override;
     qint64 maxCapacity() const override;
-    qint64 maxLabelLength() const override;
+    int maxLabelLength() const override;
     SupportTool supportToolName() const override;
     bool supportToolFound() const override;
